perf(login): trim bundle for /login (route group, no shared providers)
Lighthouse mobile on /login was 0.82 perf with 5.0s LCP because the route
shipped the same client tree as / — NextAuth SessionProvider, the
AppearanceProvider, sonner Toaster, and ServiceWorkerRegistration were all
mounted at the root layout, dragging a 224 KB shared chunk onto a page
that is just an email + password form.

- Move /login into a new (auth) route group with its own minimal layout.
- Extract the heavy providers into src/components/layout/AppShell.tsx,
  lazy-imported via next/dynamic so the login bundle never includes them.
- Set x-pathname in proxy.ts and read it in the root layout so we render
  <AppShell> only on authed routes; /login renders {children} bare.

Lighthouse mobile (simulate, headless Chrome):
  perf 0.82 -> 0.97
  LCP  5.0s -> 2.6s
  TTI  5.0s -> 2.9s
  FCP/TBT/CLS unchanged.

End-to-end credentials login verified: POST /api/auth/callback/credentials
issues authjs.session-token cookie, GET / returns 200 authenticated, GET
/login while authed redirects to /. Reports in
docs/reports/2026-05-08-lighthouse-login-tx71/.

Closes remote-dev-tx71.
