+++ src/proxy.ts
@@ -90,7 +90,11 @@ export async function proxy(request: NextRequest) {
     return NextResponse.redirect(new URL("/", request.url));
   }
 
-  return NextResponse.next();
+  // Pass pathname to the root layout (request headers) so it can skip
+  // heavy providers on /login. See src/app/layout.tsx.
+  const requestHeaders = new Headers(request.headers);
+  requestHeaders.set("x-pathname", pathname);
+  return NextResponse.next({ request: { headers: requestHeaders } });
 }
 
 export const config = {
